+++ debug/test.php
@@ -23,6 +23,15 @@ $grid = 10985381;
 
 $rep = new \searchad\reports\ReportingRequest();
 
+$resp = new \searchad\ApiResponse();
+//$html = "<!DOCTYPE html><html><head><title>Apache Tomcat/8.5.11 - Error report</title><style type=\"text/css\">h1 {font-family:Tahoma,Arial,sans-serif;color:white;background-color:#525D76;font-size:22px;} h2 {font-family:Tahoma,Arial,sans-serif;color:white;background-color:#525D76;font-size:16px;} h3 {font-family:Tahoma,Arial,sans-serif;color:white;background-color:#525D76;font-size:14px;} body {font-family:Tahoma,Arial,sans-serif;color:black;background-color:white;} b {font-family:Tahoma,Arial,sans-serif;color:white;background-color:#525D76;} p {font-family:Tahoma,Arial,sans-serif;background:white;color:black;font-size:12px;} a {color:black;} a.name {color:black;} .line {height:1px;background-color:#525D76;border:none;}</style> </head><body><h1>HTTP Status 401 - </h1><div class=\"line\"></div><p><b>type</b> Status report</p><p><b>message</b> <u></u></p><p><b>description</b> <u>This request requires HTTP authentication.</u></p><hr class=\"line\"><h3>Apache Tomcat/8.5.11</h3></body></html>";
+//try {
+//    $resp->loadResponse($html,[]);
+//} catch(\Exception $e){
+//    var_dump($e->getMessage());
+//}
+//exit;
+
 $repParams = '{
     "startTime": "2016-01-01T00:00:00.000",
     "endTime": "2017-10-01T00:00:00.000",
@@ -35,10 +44,10 @@ $repParams = '{
 $rep->loadCertificates(__DIR__ . '/test.pem', __DIR__ . '/test.key');
 $cb = function ($params) use ($rep) {
 
-    var_dump("Callback for request mode");
-    if($rep->getRequestType() === 'R'){
-        $rep->setAllowRun(false);
-    }
+//    var_dump("Callback for request mode");
+//    if($rep->getRequestType() === 'R'){
+//        $rep->setAllowRun(false);
+//    }
 };
 
 $rep->addBeforeRequestCallback($cb, ['234']);
@@ -70,10 +79,14 @@ $resp->addCallback(function ($params) {
     var_dump("response callback");
     var_dump($params);
 }, []);
-$resp->loadResponse($rep->getRawResponse(), $rep->getCurlInfo());
+try {
+    $resp->loadResponse($rep->getRawResponse(), $rep->getCurlInfo());
+} catch(\Exception $e){
+    var_dump($e->getMessage());
+}
 
-
-var_dump(json_decode($rep->getRawResponse()), $rep->getRequestBody(true));
+//var_dump($rep->getRawResponse());
+var_dump(json_decode($resp->isHttpCodeOk()));
 //$api = new \searchad\campaign\CampaignRequest();
 
 //$api->addBeforeRequestCallback(function ($params) {